feat: break TriggerCreatePage into components
